refactor(forecast): extract market settlement loader
